# docs/openapi.yaml
openapi: 3.1.0
info:
  title: StateSet Sequencer API
  description: |
    Verifiable Event Sync (VES) sequencer for deterministic event ordering,
    state projection, and cryptographic commitments.

    ## Authentication

    All API endpoints require authentication via API key or JWT token, except:
    - `POST /v1/agents/register` (public self-service registration, served under the `/api` base URL)
    - `GET /health`
    - `GET /ready`
    - **API Key**: Pass in `X-API-Key` header
    - **JWT**: Pass in `Authorization: Bearer <token>` header

    ## VES Protocol

    The VES (Verifiable Event Sync) protocol provides:
    - Cryptographic signatures on events (Ed25519)
    - Gap-free sequence numbering
    - Merkle tree commitments
    - L2 anchoring for finality
    - Inclusion proofs for any event

    ## Legacy v1 Deprecation

    The pre-VES **legacy v1 surface is formally deprecated** (planned removal:
    2027-03-01): every legacy operation is marked `deprecated: true` below.
    New integrations must use the VES `/v1/ves/*` endpoints (REST) or the
    `stateset.sequencer.v2` gRPC service (`proto/sequencer_v2.proto`).

    Legacy REST -> current successor:
    - `POST /v1/events/ingest` -> `POST /v1/ves/events/ingest`
    - `GET /v1/events` -> `GET /v1/ves/events`
    - `GET /v1/head` -> `GET /v1/ves/head`
    - `/v1/commitments*` -> `/v1/ves/commitments*`
    - `/v1/proofs*` -> `/v1/ves/proofs*`
    - `/v1/entities/*` -> `/v1/ves/entities/*`
...

...
  version: 0.8.6
...
servers:
...
security:
  - apiKey: []
  - bearerAuth: []
paths:
...
  /v1/ves/proofs/{sequence_number}:
    get:
      summary: Get inclusion proof
      description: Get a Merkle inclusion proof for an event at a specific sequence number
      operationId: getVesInclusionProof
      tags: [VES, Proofs]
      parameters:
        - name: sequence_number
          in: path
          required: true
          description: The sequence number of the event
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/storeId'
        - $ref: '#/components/parameters/batchIdQuery'
      responses:
        '200':
          description: Inclusion proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerkleProof'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    tenantId:
      name: tenant_id
      in: query
      required: true
      description: Tenant identifier
      schema:
        type: string
        format: uuid
    storeId:
      name: store_id
      in: query
      required: true
      description: Store identifier
      schema:
        type: string
        format: uuid
...
    batchIdQuery:
      name: batch_id
      in: query
      required: true
      description: Batch/commitment identifier
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      required:
        - error
        - code
        - message
      properties:
        error:
          type: string
          description: Error type
        code:
          type: integer
          description: Numeric error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        request_id:
          type: string
          description: Request correlation ID
        docs_url:
          type: string
          format: uri
          description: Link to error documentation
...
    MerkleProof:
      type: object
      properties:
        leaf_hash:
          type: string
          description: Hex-encoded leaf hash
        leaf_index:
          type: integer
        proof_path:
          type: array
          items:
            type: string
          description: Array of hex-encoded proof hashes
        merkle_root:
          type: string
          description: Hex-encoded Merkle root
        batch_id:
          type: string
          format: uuid
...